% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_box.R
\name{zero_box}
\alias{zero_box}
\title{zero_box}
\usage{
zero_box(data, group, value)
}
\arguments{
\item{data}{Input data}

\item{group}{Group variable in the data}

\item{value}{Numeric variable in the data}
}
\description{
zero_box
}
\examples{
zero_box(data=zip, group="Trt", value="Resp")

}
