% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_hist.R
\name{zero_hist}
\alias{zero_hist}
\title{zero_hist}
\usage{
zero_hist(data, group, value)
}
\arguments{
\item{data}{Input data}

\item{group}{Group variable in the data}

\item{value}{Numeric variable in the data}
}
\description{
zero_hist
}
\examples{
zero_hist(data=zip, group="Trt", value="Resp")

}
